+++ CHANGELOG.md
@@ -10,7 +10,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * * Allow combining sets of processed data
 
 * Charting
-* * Line graphs
+* * Line graph x axis label position
+* * Line graph quantitative axis
+* * Clean up labels of groups of continuous data
+* * Allow customisation of how labels are displayed
 
 ### Added